Add board-nk3xn and board-nk3am features
This patch adds support for the NK3 protypes.  board-nk3xn can be used
for the NK3AN (= USB A NFC) and the NK3CN (= USB C NFC), and board-nk3am
can be used for the NK3AM (= USB A Mini).  Currently, board-nk3am is
only an alias for the board-solo2 feature as the solo2 code also works
for the NK3AM.

It also makes it easier to select a board when using the Makefile by
adding the BOARD variable, defaulting to nk3xn.